refactor: calcListNum.js 폴더 변경 - common 폴더로 위치 변경
